LF-5215 Move functions to util file to fix animalUtils.test
