+++ packages/webapp/src/containers/Animals/utils.ts
@@ -13,7 +13,14 @@
  *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
  */
 import i18n from '../../locales/i18n';
-import { Animal, AnimalBatch } from '../../store/api/types';
+import {
+  Animal,
+  AnimalBatch,
+  CustomAnimalBreed,
+  CustomAnimalType,
+  DefaultAnimalBreed,
+  DefaultAnimalType,
+} from '../../store/api/types';
 
 const { t } = i18n;
 
@@ -29,4 +36,53 @@ export const chooseIdentification = (animalOrBatch: Animal | AnimalBatch) => {
     return animalOrBatch.name;
   }
   return `${t('ANIMAL.ANIMAL_ID')}${animalOrBatch.internal_identifier}`;
+};
+
+type hasId = {
+  id: number;
+  [key: string]: any;
+};
+
+const getProperty = (arr: hasId[] | undefined, id: number | null, key: string) => {
+  return arr?.find((el) => el.id === id)?.[key] || null;
+};
+
+const getAnimalTypeLabel = (key: string) => {
+  return t(`TYPE.${key}`, { ns: 'animal' });
+};
+
+const getAnimalBreedLabel = (key: string) => {
+  return t(`BREED.${key}`, { ns: 'animal' });
+};
+
+export const chooseAnimalTypeLabel = (
+  animalOrBatch: Animal | AnimalBatch,
+  defaultAnimalTypes: DefaultAnimalType[],
+  customAnimalTypes: CustomAnimalType[],
+) => {
+  if (animalOrBatch.default_type_id) {
+    return getAnimalTypeLabel(
+      getProperty(defaultAnimalTypes, animalOrBatch.default_type_id, 'key'),
+    );
+  } else if (animalOrBatch.custom_type_id) {
+    return getProperty(customAnimalTypes, animalOrBatch.custom_type_id, 'type');
+  } else {
+    return null;
+  }
+};
+
+export const chooseAnimalBreedLabel = (
+  animalOrBatch: Animal | AnimalBatch,
+  defaultAnimalBreeds: DefaultAnimalBreed[],
+  customAnimalBreeds: CustomAnimalBreed[],
+) => {
+  if (animalOrBatch.default_breed_id) {
+    return getAnimalBreedLabel(
+      getProperty(defaultAnimalBreeds, animalOrBatch.default_breed_id, 'key'),
+    );
+  } else if (animalOrBatch.custom_breed_id) {
+    return getProperty(customAnimalBreeds, animalOrBatch.custom_breed_id, 'breed');
+  } else {
+    return null;
+  }
 };
